Fixed culture overflow in calc code for culture decay pointed out by sp00n (Civfanatic nick). Thanks sp00n!

+++ Sources/CvPlot.cpp
@@ -10267,7 +10267,7 @@ void CvPlot::doCulture()
 
 			if (getCulture(ePlayerX) > 0 && getCultureRateThisTurn(ePlayerX) < 1 && (!getPlotCity() || getOwner() != ePlayerX))
 			{
-				setCulture(ePlayerX, getCulture(ePlayerX) * (1000 - decayPermille) / 1000, false, false, true);
+				setCulture(ePlayerX, static_cast<int>(int64_t(getCulture(ePlayerX) * (1000 - decayPermille) / 1000)), false, false, true);
 			}
 		}
 	}
